# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "AppController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/polls": {
      "get": {
        "operationId": "PollsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "polls"
        ]
      },
      "post": {
        "operationId": "PollsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePollDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollDetailDto"
                }
              }
            }
          }
        },
        "tags": [
          "polls"
        ]
      }
    },
    "/polls/{id}": {
      "get": {
        "operationId": "PollsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollDetailDto"
                }
              }
            }
          }
        },
        "tags": [
          "polls"
        ]
      },
      "patch": {
        "operationId": "PollsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePollDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollDetailDto"
                }
              }
            }
          }
        },
        "tags": [
          "polls"
        ]
      },
      "delete": {
        "operationId": "PollsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "polls"
        ]
      }
    },
    "/polls/{id}/active": {
      "patch": {
        "operationId": "PollsController_toggleActive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleActiveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollSummaryDto"
                }
              }
            }
          }
        },
        "tags": [
          "polls"
        ]
      }
    },
    "/admin/polls": {
      "get": {
        "operationId": "AdminPollsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-polls"
        ]
      }
    },
    "/admin/polls/{id}": {
      "get": {
        "operationId": "AdminPollsController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollDetailDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-polls"
        ]
      },
      "patch": {
        "operationId": "AdminPollsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePollDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollDetailDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-polls"
        ]
      },
      "delete": {
        "operationId": "AdminPollsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "admin-polls"
        ]
      }
    },
    "/admin/polls/{id}/active": {
      "patch": {
        "operationId": "AdminPollsController_toggleActive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleActiveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollSummaryDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-polls"
        ]
      }
    },
    "/public/polls/{slug}": {
      "get": {
        "operationId": "ResponsesController_getPublic",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPollDto"
                }
              }
            }
          }
        },
        "tags": [
          "public"
        ]
      }
    },
    "/public/polls/{slug}/responses": {
      "post": {
        "operationId": "ResponsesController_submit",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitResponseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "public"
        ]
      }
    },
    "/polls/{id}/analytics": {
      "get": {
        "operationId": "AnalyticsController_getOwnerAnalytics",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerAnalyticsDto"
                }
              }
            }
          }
        },
        "tags": [
          "analytics"
        ]
      }
    },
    "/admin/polls/{id}/analytics": {
      "get": {
        "operationId": "AnalyticsController_getAdminPollAnalytics",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerAnalyticsDto"
                }
              }
            }
          }
        },
        "tags": [
          "analytics"
        ]
      }
    },
    "/admin/analytics": {
      "get": {
        "operationId": "AnalyticsController_getSystemAnalytics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemAnalyticsDto"
                }
              }
            }
          }
        },
        "tags": [
          "analytics"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "UsersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users/{id}/role": {
      "patch": {
        "operationId": "UsersController_changeRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummaryDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users/bulk-delete": {
      "post": {
        "operationId": "UsersController_bulkDelete",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users/export.csv": {
      "get": {
        "operationId": "UsersController_exportCsv",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-users"
        ]
      }
    }
  },
  "info": {
    "title": "Survey App API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "alice@example.com"
          },
          "name": {
            "type": "string",
            "example": "Alice Example"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "name",
          "password"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "PollSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "responseCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "visibility",
          "isActive",
          "responseCount",
          "createdAt"
        ]
      },
      "PollListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PollSummaryDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "OptionInputDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          }
        },
        "required": [
          "text"
        ]
      },
      "QuestionInputDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "SINGLE_CHOICE",
              "MULTIPLE_CHOICE",
              "TEXT"
            ]
          },
          "text": {
            "type": "string"
          },
          "isRequired": {
            "type": "boolean",
            "default": false
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OptionInputDto"
            }
          }
        },
        "required": [
          "type",
          "text",
          "isRequired",
          "options"
        ]
      },
      "CreatePollDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ],
            "default": "PRIVATE"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionInputDto"
            }
          }
        },
        "required": [
          "title",
          "visibility",
          "isActive",
          "questions"
        ]
      },
      "OptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "text",
          "order"
        ]
      },
      "QuestionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "SINGLE_CHOICE",
              "MULTIPLE_CHOICE",
              "TEXT"
            ]
          },
          "text": {
            "type": "string"
          },
          "isRequired": {
            "type": "boolean"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OptionDto"
            }
          }
        },
        "required": [
          "id",
          "order",
          "type",
          "text",
          "isRequired",
          "options"
        ]
      },
      "PollDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "responseCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionDto"
            }
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "visibility",
          "isActive",
          "responseCount",
          "createdAt",
          "questions"
        ]
      },
      "UpdatePollDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ],
            "default": "PRIVATE"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionInputDto"
            }
          }
        },
        "required": [
          "title",
          "visibility",
          "isActive",
          "questions"
        ]
      },
      "ToggleActiveDto": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "isActive"
        ]
      },
      "PublicPollDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "closed": {
            "type": "boolean"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "closed",
          "questions"
        ]
      },
      "AnswerInputDto": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string"
          },
          "optionIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "textValue": {
            "type": "string"
          }
        },
        "required": [
          "questionId"
        ]
      },
      "SubmitResponseDto": {
        "type": "object",
        "properties": {
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerInputDto"
            }
          }
        },
        "required": [
          "answers"
        ]
      },
      "SubmitResultDto": {
        "type": "object",
        "properties": {
          "submittedAt": {
            "type": "string"
          }
        },
        "required": [
          "submittedAt"
        ]
      },
      "OptionAggregateDto": {
        "type": "object",
        "properties": {
          "optionId": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "optionId",
          "text",
          "order",
          "count"
        ]
      },
      "QuestionAggregateDto": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "SINGLE_CHOICE",
              "MULTIPLE_CHOICE",
              "TEXT"
            ]
          },
          "answerCount": {
            "type": "number"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OptionAggregateDto"
            }
          },
          "textAnswerCount": {
            "type": "number",
            "description": "Number of distinct text answers, for TEXT questions only"
          }
        },
        "required": [
          "questionId",
          "text",
          "order",
          "type",
          "answerCount",
          "options"
        ]
      },
      "OwnerAnalyticsDto": {
        "type": "object",
        "properties": {
          "pollId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "totalResponses": {
            "type": "number"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionAggregateDto"
            }
          }
        },
        "required": [
          "pollId",
          "title",
          "totalResponses",
          "questions"
        ]
      },
      "SystemAnalyticsDto": {
        "type": "object",
        "properties": {
          "totalUsers": {
            "type": "number"
          },
          "totalAdmins": {
            "type": "number"
          },
          "totalPolls": {
            "type": "number"
          },
          "activePolls": {
            "type": "number"
          },
          "totalResponses": {
            "type": "number"
          }
        },
        "required": [
          "totalUsers",
          "totalAdmins",
          "totalPolls",
          "activePolls",
          "totalResponses"
        ]
      },
      "UserSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "createdAt"
        ]
      },
      "UserListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSummaryDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "ChangeRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
      "BulkDeleteDto": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
      "BulkDeleteResultDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number"
          }
        },
        "required": [
          "count"
        ]
      }
    }
  }
}